templatized header/footer and builds
