+++ src/main/templates/facilities.hbs
@@ -1,44 +1,8 @@
 <!DOCTYPE html>
 <html lang="en">
-
-<head>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-  <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css"
-    integrity="sha384-Vkoo8x4CGsO3+Hhxv8T/Q5PaXtkKtu6ug5TOeNV6gBiFeWPGFN9MuhOf23Q9Ifjh" crossorigin="anonymous">
-  <link rel="stylesheet" type="text/css" href="mobile.css">
-  <title>ISDCF Facilities</title>
-  <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN" crossorigin="anonymous"></script>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js" integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q" crossorigin="anonymous"></script>
-  <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js" integrity="sha384-JZR6Spejh4U02d8jOt6vLEHfe/JQGiRRSQQxSfFWpi1MquVdAyjUar5+76PVCmYl" crossorigin="anonymous"></script>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.2.1/jquery.min.js" crossorigin="anonymous"></script>
-  <script type="application/javascript" src="features.js"></script>
-</head>
-
-<body>
-  <nav class="navbar navbar-expand-md navbar-light fixed-top bg-light d-print-none d-flex">
-    <a class="navbar-brand" href="#">ISDCF Facilities</a>
-    <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
-      <span class="navbar-toggler-icon"></span>
-    </button>
-    <div class="collapse navbar-collapse" id="navbarNav">
-      <ul class="navbar-nav mr-auto">
-        <li class="nav-item">
-          <a id="json-feed" target="_blank" class="nav-link" href="http://registry.isdcf.com/facilities">JSON Feed</a>
-        </li>
-        <li class="nav-item">
-          <a id="pdf-download" target="_blank" class="nav-link" title="Download as PDF" href="{{pdf_path}}">Download PDF</a>
-        </li>
-      </ul>
-        <form id="search-form" class="form-inline">
-          <span class="align-middle"><input id="search" class="form-control mb-2" type="search" placeholder="Search Facilities" aria-label="Search">
-          <input type="button" class="btn btn-secondary btn-sm mb-2" value="x" id="clear-filter"/></span>
-        </form>
-    </div>
-  </nav>
-  <a href="#" id="toTopBtn" class="cd-top text-replace js-cd-top cd-top--is-visible cd-top--fade-out" data-abc="true"></a>
-  <a class="anchor" id="top" href="#top"></a>
-  <div class="d-none d-print-block"><h4>ISDCF Facilities</h4></div>
+<header>
+  {{>header}}
+</header>
   <table class="table table-striped table-small table-hover">
     <thead class="thead-light">
       <tr>
@@ -48,7 +12,7 @@
     </thead>
     <tbody id="searchTable">
       {{#data}}
-      <tr id="facility-{{code}}">
+      <tr id="{{idType}}-{{code}}">
         <td class="code"><a class="anchor" id="{{code}}" href="#{{code}}"></a><code>{{code}}</code></td>
         <td><span class="description">{{description}}</span>
           {{#obsolete}}
@@ -57,7 +21,7 @@
               <span>Replaced by:</span>
             {{/if}}
             {{#obsoletedBy}}
-              <a class="facility-link" id="clear-filter" href="#{{this}}">{{this}}</a>
+              <a class="{{idType}}-link" id="clear-filter" href="#{{this}}">{{this}}</a>
             {{/obsoletedBy}}
           {{/obsolete}}
         </td>
@@ -66,12 +30,7 @@
     </tbody>
   </table>
   <footer>
-    <small class="text-muted">
-      <p class="text-center">Registry version: <code>{{version}}</code></p>
-      <p class="text-center">Site version: <code>{{site_version}}</code></p>
-      <p class="text-center">Generated on {{date}}</p>
-    </small>
+    {{>footer}}
   </footer>
 </body>
-
 </html>
