+++ src/main/templates/partials/footer.hbs
@@ -1,0 +1,5 @@
+    <small class="text-muted">
+      <p class="text-center">Registry version: <code>{{version}}</code></p>
+      <p class="text-center">Site version: <code>{{site_version}}</code></p>
+      <p class="text-center">Generated on {{date}}</p>
+    </small>
